Add more tests.
Change song.timeStep() to use .bind(this) instead of passing
in the song object. Fix bug discovered in testing that
didn't keep the song position fully up to date when stopping.
It was liable to get up to `song.timeStepMs` ms behind
depending on how long it was stopped after the last time step.